requirements.txt
easyocr
imutils
pytesseract
pandas
ultralytics
opencv-python
openpyxl
pyserial